faq-bot/pyproject.toml
[project]
name = "faq-bot"
version = "0.1.0"
description = "faq-bot"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "async-lru>=2.0.5",
    "beautifulsoup4>=4.13.4",
    "httpx>=0.28.1",
    "nb-cli>=1.4.2",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-alconna>=0.57.6", # required by nonebot-plugin-treehelp
    "nonebot-plugin-status>=0.9.0",
    "nonebot-plugin-treehelp>=0.5.0",
    "nonebot2[fastapi]>=2.4.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "nonebot-adapter-console>=0.6.0",
    "pip>=25.1.1",
]

[tool.nonebot]
adapters = [
    # Uncomment the following line for developing.
    # { name = "Console", module_name = "nonebot.adapters.console" },
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
plugins = [
    "nonebot_plugin_status",
    "nonebot_plugin_treehelp",
]
plugin_dirs = ["faq_bot/plugins"]
builtin_plugins = ["echo"]

[tool.ruff.lint]
extend-select = ["I", "TYP", "UP"]
